start the server bgthread
The server thread does nothing yet, but at least the pthread plumbering
is ready-ish. For an easier integration, it is actually a "bgthread" (a
pthread with a workspace), which is maybe irrelevant...

It seems that the object destructors are not implemented yet in the VCL,
so the cleanup code hasn't actually be tested.